+++ test/test_representation.clj
@@ -33,8 +33,7 @@
 (facts "Can produce representation from java.util.Map"
        (let [entity (doto (java.util.TreeMap.)
                       (.put :foo "bar")
-                      (.put :baz "qux"))
-             expected-clj (into (sorted-map) entity)]
+                      (.put :baz "qux"))]
          (tabular "Various media types are supported"
                   (as-response entity {:representation {:media-type ?media-type :charset "UTF-8"}})
                   => {:body ?body :headers { "Content-Type" (str ?media-type ";charset=UTF-8")}}
@@ -47,8 +46,8 @@
                                      "<tr><th>foo</th><td>bar</td></tr>"
                                      "</tbody></table></div>")
                   "application/json" (clojure.data.json/write-str entity)
-                  "application/clojure" (pr-str-dup expected-clj)
-                  "application/edn" (pr-str expected-clj))))
+                  "application/clojure" (pr-str-dup entity)
+                  "application/edn" (pr-str entity))))
 
 (facts "Can produce representations from a seq of maps"
        (let [entity [(sorted-map :foo 1 :bar 2) (sorted-map :foo 2 :bar 3)]]